Support for position tracking in bookmarks (destructive version).
This adds support for recording page positions along with page numbers in the
same section of the bookmarks key file. It is *not* backwards compatible with
upstream and using both versions *will* lead to corruption of the bookmarks
file. Keeping here just for the record, usage of this code is discouraged.